FIX: Root Jail rejected all requests with 'Access denied'
The comparison between the root path and the filepath did reject all
files. (At least if you .server('.') ). Fixed by resolving each path to
an absolute path with path.resolve() and comparing absolute paths. In
addition, this seems to me to be a saver comparison. Possibly reduces
the cache memory footprint as the key is now the full path. (Is only
true if previously multiple relative paths lead to the same absolute
path.)