always use options.headers to set Server header
inside finish, instead of setting Server header to the serverInfo variable, set it to the options.header.Server value, which is set in the constructor function and defaults to the serverInfo variable.

on 404 errors, the server would always report node static as the server even if the client code passed in a different Server header in the options param.