Fix for requests having both If-None-Match && If-Modified-Since
Previously, requests having both the If-None-Match and If-Modified-Since headers
would be serviced with a 304 response if either condition were to pass. This is
not correct per RFC 2616, section 14.26 (see below).

Also, added a new test to verify behavior with If-None-Match condition fails,
but If-Modified-Since condition passes.

From http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.26 :

"If none of the entity tags match, then the server MAY perform the requested
 method as if the If-None-Match header field did not exist, but MUST also ignore
 any If-Modified-Since header field(s) in the request. That is, if no entity
 tags match, then the server MUST NOT return a 304 (Not Modified) response."