Send custom headers even on a 304 Not Modified case
When a 304 (Not Modified) response is sent, user defined headers are not sent. Only default headers are.
In my case scenario, I need to update the "Set-Cookie" header to extend it every time a user accesses a page. So, even though the page has not changed, therefore generating a 304 response, I still need to send "Set-Cookie" header with a new expiration timestamp.